Introduced script to remove local bashrc and pash profile, and replace them with a hardlink to the version in this repo, what will let them update automatically when pulling the remote. 	modified:   linkBashrc.sh

+++ linkBashrc.sh
@@ -1,6 +1,18 @@
 #!/bin/bash
-rm ~/.bashrc
-ln .bashrc ~/.bashrc
+function ask_yes_or_no() {
+    read -p "$1 ([y]es or [N]o): "
+    case $(echo $REPLY | tr '[A-Z]' '[a-z]') in
+        y|yes) echo "yes" ;;
+        *)     echo "no" ;;
+    esac
+}
 
-rm ~/.bash_profile
-ln .bash_profile ~/.bash_profile
+if [[ "yes" == $(ask_yes_or_no "REMOVE ~/.bashrc and ~/.bash_profile and replace by link to files in this repo?") ]]
+         then
+
+	rm ~/.bashrc
+	ln .bashrc ~/.bashrc
+
+	rm ~/.bash_profile
+	ln .bash_profile ~/.bash_profile
+	fi
